+++ go/falcon-core/generic/listfarraydouble/testdata.go
@@ -6,7 +6,7 @@ import (
 	"github.com/falcon-autotuning/falcon-core-libs/go/falcon-core/generic/farraydouble"
 )
 
-func mustFArrayDouble(data []float64, shape []int) *farraydouble.Handle {
+func mustFArrayDouble(data []float64, shape []uint64) *farraydouble.Handle {
 	h, err := farraydouble.FromData(data, shape)
 	if err != nil {
 		panic(fmt.Errorf("failed to create FArray: %v", err))
@@ -18,10 +18,10 @@ func mustFArrayDouble(data []float64, shape []int) *farraydouble.Handle {
 
 var (
 	defaultListData = []*farraydouble.Handle{
-		mustFArrayDouble([]float64{1.0, 2.0, 3.0, 4.0}, []int{2, 2}),
-		mustFArrayDouble([]float64{5.0, 2.0, 3.1, 4.0}, []int{2, 2}),
+		mustFArrayDouble([]float64{1.0, 2.0, 3.0, 4.0}, []uint64{2, 2}),
+		mustFArrayDouble([]float64{5.0, 2.0, 3.1, 4.0}, []uint64{2, 2}),
 	}
 	otherListData = []*farraydouble.Handle{
-		mustFArrayDouble([]float64{5.0, 2.0, 3.1, 4.0, 2.0, 5.0, 6.09, 4.3, 2.3}, []int{3, 3}),
+		mustFArrayDouble([]float64{5.0, 2.0, 3.1, 4.0, 2.0, 5.0, 6.09, 4.3, 2.3}, []uint64{3, 3}),
 	}
 )
