+++ go/falcon-core/math/arrays/controlarray/controlArray_test.go
@@ -9,7 +9,7 @@ import (
 )
 
 var (
-	defaultShape = []int{2, 2}
+	defaultShape = []uint64{2, 2}
 	defaultData  = []float64{1., .2, 1., .2}
 	defaultVal   = float64(5.2)
 )
@@ -63,7 +63,7 @@ func TestControlArray_ShapeDimensionSize(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Dimension failed: %v", err)
 	}
-	if dim != uint32(len(defaultShape)) {
+	if dim != uint64(len(defaultShape)) {
 		t.Errorf("Expected dimension %d, got %d", len(defaultShape), dim)
 	}
 	sz, err := a.Size()
