+++ go/falcon-core/math/arrays/controlarray1d/controlArray1D_test.go
@@ -9,7 +9,7 @@ import (
 )
 
 var (
-	defaultShape = []int{4}
+	defaultShape = []uint64{4}
 	defaultData  = []float64{1., 2., 3., 4.}
 	defaultVal   = float64(2.0)
 )
@@ -63,7 +63,7 @@ func TestControlArray1D_ShapeDimensionSize(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Dimension failed: %v", err)
 	}
-	if dim != uint32(len(defaultShape)) {
+	if dim != uint64(len(defaultShape)) {
 		t.Errorf("Expected dimension %d, got %d", len(defaultShape), dim)
 	}
 	sz, err := a.Size()
