+++ go/falcon-core/math/axescontrolarray1d/testdata.go
@@ -4,7 +4,7 @@ import (
 	"github.com/falcon-autotuning/falcon-core-libs/go/falcon-core/math/arrays/controlarray1d"
 )
 
-func mustControlArray(data []float64, shape []int) *controlarray1d.Handle {
+func mustControlArray(data []float64, shape []uint64) *controlarray1d.Handle {
 	h, err := controlarray1d.FromData(data, shape)
 	if err != nil {
 		panic(err)
@@ -14,10 +14,10 @@ func mustControlArray(data []float64, shape []int) *controlarray1d.Handle {
 
 var (
 	defaultAxesData = []*controlarray1d.Handle{
-		mustControlArray([]float64{0.0, 1.0}, []int{2}),
-		mustControlArray([]float64{0.0, 1.0, 2.0, 3.0}, []int{4}),
+		mustControlArray([]float64{0.0, 1.0}, []uint64{2}),
+		mustControlArray([]float64{0.0, 1.0, 2.0, 3.0}, []uint64{4}),
 	}
 	otherAxesData = []*controlarray1d.Handle{
-		mustControlArray([]float64{0.5, 1.0, 2.0, 3.0}, []int{4}),
+		mustControlArray([]float64{0.5, 1.0, 2.0, 3.0}, []uint64{4}),
 	}
 )
