+++ go/falcon-core/math/axesdouble/axesDouble_test.go
@@ -34,7 +34,7 @@ func TestAxesDouble_NewEmptyAndPushBack(t *testing.T) {
 			t.Fatalf("PushBack failed at %d: %v", i, err)
 		}
 	}
-	if sz, err := l.Size(); err != nil || sz != uint32(len(fixtureAxesData())) {
+	if sz, err := l.Size(); err != nil || sz != uint64(len(fixtureAxesData())) {
 		t.Errorf("Expected size %d, got %d, err: %v", len(fixtureAxesData()), sz, err)
 	}
 }
