+++ go/falcon-core/physics/config/core/adjacency/adjacency_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 var (
-	defaultShape = []int{2, 2}
+	defaultShape = []uint64{2, 2}
 	defaultData  = []int32{1, 2, 3, 4}
 )
 
@@ -75,7 +75,7 @@ func TestAdjacency_ShapeDimensionSize(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Dimension failed: %v", err)
 	}
-	if dim != uint32(len(defaultShape)) {
+	if dim != uint64(len(defaultShape)) {
 		t.Errorf("Expected dimension %d, got %d", len(defaultShape), dim)
 	}
 	sz, err := a.Size()
